pyproject.toml
[project]
name = "visagent"
version = "0.1.0"
description = "Agentic AI + Vision + RAG project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.10.0",
    "chromadb>=1.0.20",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "groq>=0.31.0",
    "langchain>=0.3.27",
    "langgraph>=0.6.5",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pypdf>=6.0.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.48.1",
    "timm>=1.0.19",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.55.2",
    "uvicorn>=0.35.0",
]
